THRIFT-411. cpp: Make TNonblockingServer::connectionStackLimit_ accessors use the right type
git-svn-id: https://svn.apache.org/repos/asf/incubator/thrift/trunk@761481 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/cpp/src/server/TNonblockingServer.h
@@ -171,7 +171,7 @@ class TNonblockingServer : public TServer {
    *
    * @return the current limit on TConnection pool size.
    */
-  int getConnectionStackLimit() const {
+  size_t getConnectionStackLimit() const {
     return connectionStackLimit_;
   }
 
@@ -180,7 +180,7 @@ class TNonblockingServer : public TServer {
    *
    * @param sz the new limit for TConnection pool size.
    */
-  void setConnectionStackLimit(int sz) {
+  void setConnectionStackLimit(size_t sz) {
     connectionStackLimit_ = sz;
   }
